requirements.txt
requests
pandas
python-dotenv
snowflake-connector-python
schedule
streamlit   # Optional, only needed for dashboard
